+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * Made `favored()` S3 and added `favored.elo.glm()`. (#38)
 
+* Fixed a bug with `adjust()` variables not getting subsetted correctly with `na.action` in `model.frame()`. (#40)
+
 * Added `is.na.elo.adjust()` to test for NAs in the adjustment vector. (#41)
 
 # elo v1.1.0